refactor: Removed unused code and refactored TeamDetail component
